Enhance session management; add cancel functionality to confirmation modal, implement card review logic, and remove deprecated new-words page
